Create an abstract version of Sdl2Window
